:bug: Fixed bug where fade-in wouldn't work on page load
When doing SSR the class that applies the transition would already be part of the html. This caused the transition to not work, so I had to find a way to have the SSR page without that class.

The solution is to use a prop that is true only in the client side, using componentDidMount

Reference: https://stackoverflow.com/questions/52491619/